Rename file ReportMappingUtil.ts to report-mapping-util.ts
