+++ .npmrc
@@ -1,0 +1,3 @@
+registry=https://nexus.imatia.com/repository/npm-group/
+//nexus.imatia.com/repository/npm-group/:username=patricia.martinez
+//nexus.imatia.com/repository/npm-group/:_password=cG10LmltYXRpYS4yMDI1
