+++ projects/ontimize-web-ngx-report/src/lib/services/o-report-response.adapter.ts
@@ -2,7 +2,7 @@ import { HttpErrorResponse, HttpResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { OntimizeServiceResponse, OntimizeServiceResponseAdapter } from 'ontimize-web-ngx';
 
-import { OReportMappingUtils } from '../util/ReportMappingUtils';
+import { OReportMappingUtils } from '../util/report-mapping-utils';
 
 @Injectable()
 export class OReportResponseAdapter implements OntimizeServiceResponseAdapter {
